# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for user avatar management",
        "title": "GophProfile",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/avatars": {
            "post": {
                "description": "Uploading new user avatar with thumbnails (100x100, 300x300)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "avatars"
                ],
                "summary": "Uploading new user avatar",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "file binary data",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SaveAvatarResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "413": {
                        "description": "Request Entity Too Large"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/avatars/{avatar_id}": {
            "get": {
                "description": "returns user avatar by id (binary)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/png",
                    "image/jpeg",
                    "image/webp"
                ],
                "tags": [
                    "avatars"
                ],
                "summary": "Providing user avatar by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "avatar id",
                        "name": "avatar_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "original",
                            "100x100",
                            "300x300"
                        ],
                        "type": "string",
                        "description": "size",
                        "name": "size",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Binary image data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "removes user avatar by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "avatars"
                ],
                "summary": "Removing user avatar by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "avatar id",
                        "name": "avatar_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/avatars/{avatar_id}/metadata": {
            "get": {
                "description": "returns user avatar metadata by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "avatars"
                ],
                "summary": "Providing user avatar metadata by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "avatar id",
                        "name": "avatar_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "avatar metadata",
                        "schema": {
                            "$ref": "#/definitions/models.GetAvatarMetadata"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check status of minio, database and rabbitmq.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "App services health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "Each"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/models.HealthResponse"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/avatar": {
            "get": {
                "description": "returns binary data of last user avatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/png",
                    "image/jpeg",
                    "image/webp"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Providing last user avatar",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Binary image data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "removes last user avatar",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "removes last user avatar",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/{user_id}/avatars": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Returns list of user avatars metadata",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserAvatarsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Dimensions": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "integer"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "models.GetAvatarMetadata": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dimensions": {
                    "$ref": "#/definitions/models.Dimensions"
                },
                "file_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mime_type": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "thumbnails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetThumbnailMetadata"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.GetThumbnailMetadata": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.GetUserAvatarsListResponse": {
            "type": "object",
            "properties": {
                "avatars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetAvatarMetadata"
                    }
                }
            }
        },
        "models.HealthResponse": {
            "type": "object",
            "properties": {
                "broker": {
                    "type": "boolean"
                },
                "database": {
                    "type": "boolean"
                },
                "s3": {
                    "type": "boolean"
                }
            }
        },
        "models.SaveAvatarResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}